feat: 홈화면 - 작성중인 회고 swiper 적용 (#499)
* feat: #496 작성중인 회고 - swiper 적용 및 그라데이션 적용

* remove: #496 컴포넌트 제거

* chore: #496 desktop 디렉토리로 묶기
