+++ apps/web/src/app/desktop/home/HomePage.tsx
@@ -1,7 +1,7 @@
 import { css } from "@emotion/react";
 
-import HomePageHeader from "@/component/home/HomePageHeader";
-import InProgressRetrospectsWrapper from "@/component/home/InProgressRetrospectsWrapper";
+import HomePageHeader from "@/component/home/desktop/HomePageHeader";
+import InProgressRetrospectsWrapper from "@/component/home/desktop/InProgressRetrospectsWrapper";
 
 export function HomePage() {
   return (
